Add stock/director

+++ src/main/java/nl/inholland/apidemo/model/MovieNotFoundException.java
@@ -2,7 +2,7 @@ package nl.inholland.apidemo.model;
 
 public class MovieNotFoundException extends RuntimeException {
 
-  public MovieNotFoundException(int id) {
+  public MovieNotFoundException(long id) {
     super("Could not find the specified movie.;" + id);
   }
 }
